Update #10 complete edit admin

+++ src/CONTROLLER/ct_Admin.java
@@ -32,6 +32,7 @@ import java.io.File;
 import java.io.IOException;
 import java.net.URL;
 import java.sql.Timestamp;
+import java.text.ParseException;
 import java.util.List;
 import java.util.Optional;
 import java.util.ResourceBundle;
@@ -55,30 +56,49 @@ public class ct_Admin implements Initializable {
     AdminEntity SelectedAd = new AdminEntity();
     TaikhoanEntity SelectedAdacc = new TaikhoanEntity();
 
+
+    public void EditAdmin() throws IOException, ParseException {
+        if(SelectedAd.getTenAd()!=null){
+            FXMLLoader loader = new FXMLLoader(getClass().getResource("../VIEW/Form/fadd_edit_Admin.fxml"));
+            Parent root = loader.load();
+            ct_addeditAdmin ct =loader.getController();
+            ct.setInformation_Admin(SelectedAd);
+            Stage stage = new Stage();
+            stage.setScene(new Scene(root));
+            stage.initStyle(StageStyle.UTILITY);
+            if(stage.isShowing()==false){
+                stage.show();
+            }
+        }
+
+    }
+
     public void DeleteAdmin(){
         adminDao adminDao = new adminDao();
         taikhoanDao taikhoanDao = new taikhoanDao();
         System.out.println(SelectedAd.getTenAd());
+        if(SelectedAd.getTenAd()!=null){
+            Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+            alert.setTitle("Delete Admin");
+            alert.setHeaderText("Are you sure you want delete "+ SelectedAd.getTenAd());
+            Optional<ButtonType> option = alert.showAndWait();
+            if(option.get()==ButtonType.OK) {
+                SelectedAdacc = taikhoanDao.getUser_Pass(SelectedAd.getTaikhoanByUsername().getUsername());
+                System.out.println(SelectedAdacc.getUsername());
+                File image = new File("src/"+SelectedAd.getImgAd());
+                if (image.exists()&& image.delete()){
+                    System.out.println("image delete :"+image.getName());
+                }else {
+                    System.out.println("image delete fail" +image.getPath());
+                }
+                System.out.println("admin delete: "+adminDao.delData(SelectedAd));
+                System.out.println("account delete:"+taikhoanDao.delData(SelectedAdacc));
+            }else if(option.get()==ButtonType.CANCEL){
+                alert.close();
+            }else
+                alert.close();
+        }
 
-        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
-        alert.setTitle("Delete Admin");
-        alert.setHeaderText("Are you sure you want delete "+ SelectedAd.getTenAd());
-        Optional<ButtonType> option = alert.showAndWait();
-        if(option.get()==ButtonType.OK) {
-            SelectedAdacc = taikhoanDao.getUser_Pass(SelectedAd.getTaikhoanByUsername().getUsername());
-            System.out.println(SelectedAdacc.getUsername());
-            File image = new File("src/"+SelectedAd.getImgAd());
-            if (image.exists()&& image.delete()){
-                System.out.println("image delete :"+image.getName());
-            }else {
-                System.out.println("image delete fail" +image.getPath());
-            }
-            System.out.println("admin delete: "+adminDao.delData(SelectedAd));
-            System.out.println("account delete:"+taikhoanDao.delData(SelectedAdacc));
-        }else if(option.get()==ButtonType.CANCEL){
-            alert.close();
-        }else
-            alert.close();
 
 
     }
